Version 1.0.0
Completed and perfected the pipeline, removing redundancies, and made sure it all works flawlessly in PowerBI.

I removed useless fact tables and condensed them in just one, making it easier to use and more flexible in case something must be added.

Still the .csv logic is implemented; sadly, due to hardware contraints, the usage of a duckdb connector to PowerBI is impossible, thus data must be uploaded through .csv files.

Various comment lines are still up for safety but will soon be removed.